# docs/swagger.yaml
definitions:
  model.LwtArticleAdjustment:
    properties:
      article_no:
        description: ArticleNo 品类序号
        type: integer
      final_declared_value_old:
        description: FinalDeclaredValueOld 当前申报价值
        type: number
      final_declared_value_target:
        description: FinalDeclaredValueTarget 原报关单申报价值（微调的目标值）
        type: number
    type: object
  model.RequestForLwtAdjustment:
    properties:
      adjustments:
        items:
          $ref: '#/definitions/model.LwtArticleAdjustment'
        type: array
      customs_id:
        description: CustomsId 海关编号，非必填
        type: string
      lwt_filename:
        description: LwtFilename 已经生成的LWT文件名
        type: string
    type: object
  model.ResponseForLwtAdjustment:
    properties:
      customs_id:
        description: CustomsId 海关编号，非必填
        type: string
      errors:
        type: string
      has_diff:
        description: 是否依然有差异，true表示有差异，false表示没有差异
        type: boolean
      lwt_filename:
        description: LwtFilename 用于下载新的LWT文件
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
paths:
  /lwt/{filename}:
    get:
      consumes:
      - application/json
      description: get file by filename
      parameters:
      - description: LWT filename
        in: path
        name: filename
        required: true
        type: string
      - description: Download file
        in: query
        name: download
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Download excel for LWT
      tags:
      - lwt
  /lwt/adjust:
    post:
      consumes:
      - application/json
      description: 'Adjust LWT calculation parameters and regenerate LWT file.IMPORTANT:
        The customs''s LWT file must be generated first.'
      parameters:
      - description: LWT adjustment request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.RequestForLwtAdjustment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseForLwtAdjustment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResponseForLwtAdjustment'
      summary: Adjust LWT parameters
      tags:
      - lwt
swagger: "2.0"